Fix role by id struct conversion

+++ get_user_by_id.go
@@ -23,7 +23,7 @@ type GetUserByIDResult struct {
 		ID            int64             `json:"id"`
 		Status        int               `json:"status"`
 		State         int               `json:"state"`
-		RolesID       []int16           `json:"role_id"`
+		RolesID       []int64           `json:"role_id"`
 		ManagerUserID int64             `json:"manager_user_id"`
 		MemberOf      string            `json:"member_of"`
 		CustomAttrs   map[string]string `json:"custom_attributes"`
